requirements.txt
Django==5.0.2
django-environ==0.11.2
djangorestframework==3.15.1
gunicorn==21.2.0
psycopg[binary]
whitenoise==6.6.0
Pillow
djangorestframework
django-cors-headers